Youtube_rater/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
django-cors-headers
gunicorn==20.0.4